third_party/blink/web_tests/virtual/input-closes-select/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-parsing.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Input tags should parse inside select instead of closing the select
  assert_equals: expected "<input>" but got ""
Harness: the test ran to completion.
